requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
alembic==1.13.1
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
better-auth==0.0.1  # Placeholder - actual Better Auth dependency